requirements.txt
# PARA Method Template System Requirements

# Template engine
Jinja2>=3.1.0

# YAML configuration parsing
PyYAML>=6.0

# Frontmatter parsing for notes
python-frontmatter>=1.0.0

# Optional: Enhanced CLI interface
click>=8.0.0